+++ host_info.h
@@ -7,6 +7,6 @@ using std::string;
 extern int get_num_cpus();
 extern string get_ip_address(bool ipv6);
 extern string get_hostname();
-extern unsigned long get_system_memory();
+extern unsigned long long int get_system_memory();
 
 #endif
